integration with dockerised dynamo db and test refactor
