+++ docker-compose.yaml
@@ -1,0 +1,32 @@
+services:
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    depends_on:
+      - dynamodb-local
+    environment:
+      - NODE_ENV=test
+      - AWS_REGION=us-east-1
+      - DYNAMODB_ENDPOINT=http://dynamodb-local:8000
+    ports:
+      - "3000:3000" # Assuming app runs on port 3000
+
+  dynamodb-local:
+    image: amazon/dynamodb-local
+    container_name: dynamodb-local
+    ports:
+      - "8000:8000"
+
+  aws-cli:
+    image: amazon/aws-cli
+    depends_on:
+      - dynamodb-local
+    environment:
+      - AWS_ACCESS_KEY_ID=dummy
+      - AWS_SECRET_ACCESS_KEY=dummy
+      - AWS_DEFAULT_REGION=us-east-1
+    entrypoint: ""
+    volumes:
+      - ./setup-dynamodb.sh:/setup-dynamodb.sh
+    command: "/setup-dynamodb.sh"
