+++ package.json
@@ -7,7 +7,8 @@
     "scripts": {
         "build": "tsc",
         "watch": "tsc -w",
-        "test": "jest",
+        "test:unit": "jest --testPathPattern=test/unit" ,
+        "test:integration": "jest --testPathPattern=test/integration",
         "cdk": "cdk",
         "deploy": "cdk deploy --require-approval=never"
     },
