+++ test/unit/createPayment.test.ts
@@ -1,7 +1,7 @@
-import * as payments from '../src/lib/payments';
+import * as payments from '../../src/lib/payments';
 import { randomUUID } from 'crypto';
 import { APIGatewayProxyEvent } from 'aws-lambda';
-import { handler } from '../src/createPayment';
+import { handler } from '../../src/createPayment';
 
 describe('When the user requests a payment to be created', () => {
     afterEach(() => {
